Improve Support for Scout: Add support for array in newQueryForRestoration($ids)
These changes will improve support with Scout. When using models with the Searchable trait, an array is passed to the newQueryWithoutScopes(..) method. The current method in this repo only supports a value, and the following exception occurs during execution:
base64_decode() expects parameter 1 to be string, array given

The method will now support arrays with binary data and improves the out of box code in the master branch of the Laravel repo